fold AbstractDistributedTicketRegistry into AbstractTicketRegistry
