+++ cas-server-integration-hazelcast/src/main/java/org/jasig/cas/ticket/registry/HazelcastTicketRegistry.java
@@ -1,11 +1,11 @@
 package org.jasig.cas.ticket.registry;
 
-import com.hazelcast.core.HazelcastInstance;
-import com.hazelcast.core.IMap;
 import org.jasig.cas.ticket.ServiceTicket;
 import org.jasig.cas.ticket.Ticket;
 import org.jasig.cas.ticket.TicketGrantingTicket;
 
+import com.hazelcast.core.HazelcastInstance;
+import com.hazelcast.core.IMap;
 import org.springframework.beans.factory.DisposableBean;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
@@ -28,7 +28,7 @@ import java.util.concurrent.TimeUnit;
  * @since 4.1.0
  */
 @Component("hazelcastTicketRegistry")
-public class HazelcastTicketRegistry extends AbstractDistributedTicketRegistry implements DisposableBean {
+public class HazelcastTicketRegistry extends AbstractTicketRegistry implements DisposableBean {
 
     private final IMap<String, Ticket> registry;
