+++ cas-server-integration-infinispan/src/main/java/org/jasig/cas/ticket/registry/InfinispanTicketRegistry.java
@@ -1,8 +1,8 @@
 package org.jasig.cas.ticket.registry;
 
-import org.infinispan.Cache;
 import org.jasig.cas.ticket.Ticket;
 
+import org.infinispan.Cache;
 import org.springframework.stereotype.Component;
 
 import javax.annotation.Resource;
@@ -18,7 +18,7 @@ import java.util.HashSet;
  * @since 4.2.0
  */
 @Component("infinispanTicketRegistry")
-public final class InfinispanTicketRegistry extends AbstractDistributedTicketRegistry {
+public final class InfinispanTicketRegistry extends AbstractTicketRegistry {
 
     @Resource(name="infinispanTicketsCache")
     private Cache<String, Ticket> cache;
